added create post icon
