+++ src/components/home/homefeed/CreatePost.tsx
@@ -1,9 +1,10 @@
 "use client";
 import QuillEditor from "@/components/shared/QuillEditor";
 import { useCreatePostMutation } from "@/redux/api/postApi";
-import { Button, Input, Modal, message } from "antd";
+import { Button, Modal, Tooltip, message } from "antd";
 import Head from "next/head";
 import React, { useState } from "react";
+import { CiCirclePlus } from "react-icons/ci";
 
 const CreatePost: React.FC = () => {
   const [modalVisible, setModalVisible] = useState(false);
@@ -55,28 +56,24 @@ const CreatePost: React.FC = () => {
   };
 
   return (
-    <div className="p-5">
+    <div className="p-5  z-10">
       <Head>
         <link
           rel="stylesheet"
           href="https://cdnjs.cloudflare.com/ajax/libs/antd/4.18.3/antd.min.css"
         />
       </Head>
-      <div className="flex items-center mb-4 bg-white h-24 w-full p-5 rounded-2xl">
-        <img
-          src="https://via.placeholder.com/150"
-          alt="User image"
-          width={40}
-          height={40}
-          className="rounded-full"
-        />
-        <div className="ml-3 w-full">
-          <Input
-            placeholder="What you want to tell us."
-            onClick={handleInputClick}
-            className="h-16"
-          />
-        </div>
+      <div
+        className="fixed bottom-8 right-8 text-gray-600 hover:text-white rounded-full  cursor-pointer hover:bg-blue-500 transition duration-300 ease-in-out hover:rotate-360"
+        onClick={handleInputClick}
+      >
+        <Tooltip
+          title="Make a post"
+          overlayClassName="rounded-lg p-2 bg-gray-800 text-white text-sm"
+        >
+          {" "}
+          <CiCirclePlus size={80} />
+        </Tooltip>
       </div>
 
       <Modal
@@ -91,7 +88,7 @@ const CreatePost: React.FC = () => {
             key="submit"
             type="primary"
             onClick={handleCreatePost}
-            className={`mt-2 px-4 py-2 text-white rounded  bg-gray-600 hover:bg-gray-700 ml-3 ${
+            className={`mt-2 px-4 py-2 text-white rounded bg-gray-600 hover:bg-gray-700 ml-3 ${
               isLoading ? "cursor-not-allowed" : "cursor-pointer"
             }`}
           >
@@ -100,7 +97,6 @@ const CreatePost: React.FC = () => {
         ]}
       >
         <div className="my-4">
-          {/* need to chenge the editor  */}
           <QuillEditor
             valueEditor={valueEditor}
             setValueEditor={setValueEditor}
